shell: treat alarm id as name if it is not uuid like
We can specify alarm name to do alarm {show, update, delete}, but
need to explicitly use --alarm-name, which seems not so convenient.

This patch adds a check for postional argument, if it is UUID like,
we treat it as alarm id, if not, we treat it as name.

However, there is a rare case that a name can be UUID like as well,
command will fail with 404 and people will treat it as a bug, so we do
one more request just in case.

User can avoid this by specifying --name <NAME> to avoid that.

Co-Authored-By: ZhiQiang Fan <aji.zqfan@gmail.com>

Change-Id: I6218f686ea2e49a0aa4a682078522afe92b9e122